About this NodeJS boot camp
A Node.js server is meant to handle RESTful requests, mainly. We first build a basic application to process GET and POST requests. Then we will expand the app to include routing that will add functionality to the basic application. We will install several packages including Nodemon and Mongoose. In the end we will have a basic application that interacts with a No-SQL database.